refactor(web): the file-node picker's rows derive from the link targets
Unify increment #16 (ticket unify-document-pages-behind-backend-port).
Both pages built fileRefOptions inline as a third projection of the same
document rows; it is now linkTargets under the picker's field names
(lib/link-entries.ts fileRefOptions), so the three lists a document can
appear in cannot label it three different ways.

Two deliberate behavior alignments, both toward what the link picker
already did: the daemon page's file-node picker labels by DISPLAY NAME
(falling back to the path) instead of always the path, and the browser
page's rows come from the switcherOptions overlay (live-snapshot
freshness) instead of the raw list.

FileRefOption moved from DocumentPickerDialog.tsx to lib/link-entries.ts —
layer-order correctly rejected lib importing a component's type, and its
own doc says the type was filed in the wrong layer; seven importers
repointed.

Verified: link-entries 8/8 (red-first for the projection), layer-order
4/4, both pages' jsdom suites + file-seam-conformance 111/111,
canvas-ref-node real-browser 5/5, web typecheck, lint, knip clean.
