+++ apps/web/src/components/spatial-editor/CanvasContextMenu.tsx
@@ -4,6 +4,7 @@ import type { FacetRegistry } from '@kamiazya/whiteboard-facet-engine'
 import type { ClipboardFragment, SpatialCanvas, SpatialNode } from '@kamiazya/whiteboard-model'
 import { bundledFacetRegistry } from '@kamiazya/whiteboard-plugin-visual'
 import type { MutableRefObject } from 'react'
+import type { FileRefOption } from '../../lib/link-entries.js'
 import type { Point } from '../../lib/spatial/viewport.js'
 import type { ResolvedTheme } from '../../lib/theme.js'
 import type { BoxMove } from './align.js'
@@ -13,7 +14,6 @@ import { canvasMenuItems } from './context-menu-items/canvas-menu-items.js'
 import { commentMenuItems } from './context-menu-items/comment-menu-items.js'
 import { edgeMenuItems } from './context-menu-items/edge-menu-items.js'
 import { nodeMenuItems } from './context-menu-items/node-menu-items.js'
-import type { FileRefOption } from './DocumentPickerDialog.js'
 import type { GestureResult, GestureState } from './gestures.js'
 
 /** Open right-click menu: screen position (root-relative) + hit target. */
