+++ apps/web/src/components/spatial-editor/use-file-seam-scene.ts
@@ -3,8 +3,8 @@ import type { SpatialCanvas, SpatialNode } from '@kamiazya/whiteboard-model'
 import { useEffect, useMemo, useState } from 'react'
 import { createSpatialContentCache } from '../../lib/content-cache.js'
 import { composeReferenceSeam } from '../../lib/layout-worker-protocol.js'
+import type { FileRefOption } from '../../lib/link-entries.js'
 import type { ResolvedTheme } from '../../lib/theme.js'
-import type { FileRefOption } from './DocumentPickerDialog.js'
 
 /**
  * The LOD gate (embed spec v2, user decision 2026-08-08): a file node
