+++ apps/web/src/lib/link-entries.ts
@@ -69,3 +69,33 @@ export function linkTargets(
     ]
   })
 }
+
+/**
+ * The file-node picker's rows: the link targets under the field names the
+ * picker speaks (`file`/`label`). Derived rather than built inline a third
+ * time per page, so the three lists a document can appear in — link picker,
+ * switcher, file-node picker — cannot label it three different ways.
+ */
+/**
+ * A row the file-node picker offers. Filed in `lib/` because it is a pure
+ * data contract shared by the pages that build the rows and the picker that
+ * reads them — layer-order forbids `lib/` reaching into `components/` for it.
+ */
+export interface FileRefOption {
+  /** Opaque reference stored as the file node's `file` value. */
+  readonly file: string
+  /** Human-readable name shown in the picker. */
+  readonly label: string
+  /**
+   * What the referenced document IS, used to size the node the picker
+   * creates. A markdown document renders its prose inside the node, which
+   * needs room a reference card does not — at the card's height not one
+   * block fits and the body silently does not render at all. Absent keeps
+   * the card-sized default.
+   */
+  readonly kind?: DocumentKind
+}
+
+export function fileRefOptions(targets: readonly LinkTarget[]): readonly FileRefOption[] {
+  return targets.map((target) => ({ file: target.id, label: target.name, kind: target.kind }))
+}
